tidy up utils
