+++ src/remarkable-ui-embeddables/theme/formatter/formatter.utils.ts
@@ -1,31 +1,8 @@
 import { DimensionOrMeasure } from '@embeddable.com/core';
 import { DateTimeFormatter, NumberFormatter, StringFormatter } from './formatter.types';
 import { Theme } from '../theme.types';
-
-/**
- * Creates a formatter cache.
- * Cache used to prevent unnecessary (expensive) creation of formatter objects
- */
-export const cache = <Params, Formatter>(factory: (params?: Params) => Formatter) => {
-  const cache: { [key: string]: Formatter } = {};
-  const get = (params?: Params) => {
-    const key = JSON.stringify(params);
-    let formatter = cache[key];
-    if (formatter) {
-      return formatter;
-    }
-    formatter = factory(params);
-    cache[key] = formatter;
-    return formatter;
-  };
-  return get;
-};
-
-const ISO_DATE_TIME_REGEX = /^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}$/;
-
-const isValidDate = (value: string): boolean => {
-  return Boolean(value && ISO_DATE_TIME_REGEX.test(value));
-};
+import { cache } from '../../utils.ts/cache.utils';
+import { isValidISODate } from '../../utils.ts/data.utils';
 
 export type GetThemeFormatter = {
   number: (value: number | bigint, options?: Intl.NumberFormatOptions) => string;
@@ -72,7 +49,7 @@ export const getThemeFormatter = (theme: Theme): GetThemeFormatter => {
       }
 
       // Time
-      if (key.nativeType === 'time' && isValidDate(value)) {
+      if (key.nativeType === 'time' && isValidISODate(value)) {
         newValue = cachedDataDateTimeFormatter(key).format(new Date(value));
       }
